#290 Improve settings page
