+++ client/app/routes/dashboard/layout.tsx
@@ -130,19 +130,25 @@ export default function Layout() {
           title: "Organization",
           children: [
             {
-              id: "organization-general",
-              title: "General",
-              href: `/${slug}/settings/organization`,
-            },
-            {
               id: "organization-members",
               title: "Members",
-              href: `/${slug}/settings/members`,
+              href: `/${slug}/settings/organization/members`,
             },
             {
               id: "organization-teams",
               title: "Teams",
-              href: `/${slug}/settings/teams`,
+              href: `/${slug}/settings/organization/teams`,
+            },
+          ],
+        },
+        {
+          id: "cluster-group",
+          title: "Cluster",
+          children: [
+            {
+              id: "api-keys",
+              title: "API Keys",
+              href: `/${slug}/settings/cluster`,
             },
           ],
         },
