+++ client/app/routes/dashboard/settings/index.tsx
@@ -3,7 +3,7 @@ import type { Route } from "./+types/index";
 import { redirect } from "react-router";
 
 export async function loader(loaderArgs: Route.LoaderArgs) {
-  throw redirect(`/${loaderArgs.params.slug}/settings/organization`);
+  throw redirect(`/${loaderArgs.params.slug}/settings/organization/members`);
 }
 
 export default function Settings() {
