+++ client/app/routes/dashboard/settings/organization/teams/index.tsx
@@ -122,3 +122,4 @@ export default function Teams() {
     </div>
   );
 }
+
